Implemented / fixed / polished BarbarianMover script.

+++ Assets/Source/Scripts/Controllers/PlayerController.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class PlayerController : MonoBehaviour
 {	
 	public float speed;
-	public int shipMovementBoundaryX1, shipMovementBoundaryX2, shipMovementBoundaryY1, shipMovementBoundaryY2;
+	public float shipMovementBoundaryX1, shipMovementBoundaryX2, shipMovementBoundaryY1, shipMovementBoundaryY2;
 	public static float cameraDistance;
 	public static GameObject playerController;
 	ModifierCombo modifierCombo;
